Refined project files.
