+++ src/styles/chota.custom.css
@@ -2,3 +2,11 @@
     --color-primary: #333333;
     --bg-color: #f1eee8;
 }
+
+body.dark {
+    --bg-color: #000;
+    --bg-secondary-color: #131316;
+    --font-color: #f5f5f5;
+    --color-grey: #ccc;
+    --color-darkGrey: #777;
+}
